Create 'settings' module and move related code over
