+++ src/bin/show_users.rs
@@ -1,8 +1,8 @@
 extern crate schani_userinfo;
 extern crate diesel;
 
-use self::schani_userinfo::get_users;
 use schani_userinfo::db::establish_connection;
+use schani_userinfo::users::get_users;
 
 fn main() {
     let connection = establish_connection();
